fix(cli-sub-agent): propagate subtree pin through all CSA-recursion spawn paths (#1741)
Round-4 review found csa review/debate inherited the parent pin for their own
tool selection (round-1) but did not propagate it: the spawned reviewer/debater
child process never received the subtree pin env, so a nested CSA call from that
Layer-2 worker re-selected the tier default. This closes the consume/propagate
parity gap across every CSA-recursion spawn path.

- review_cmd_execute.rs + debate_cmd.rs: inject_subtree_model_pin_env into the
  reviewer/debater child env after build_execution_env (so it survives the
  round-3 funnel strip), mirroring csa run at run_cmd_attempt.rs:309.
- batch.rs, plan_cmd_exec.rs, claude_sub_agent_cmd.rs: propagate an inherited
  pin via the new propagate_inherited_subtree_pin pass-through (self-gated:
  no-op at root depth or when unpinned).
- run_cmd_model_pin.rs: shared helpers current_depth_from_env() and
  propagate_inherited_subtree_pin(); consume-path sites inject directly with
  their resolved spec. run_cmd_attempt.rs: parity doc-comment.
- mcp_server excluded (third-party MCP backend spawn, not CSA recursion).
- claude_sub_agent_cmd.rs split: extract #[cfg(test)] mod tests into sibling
  claude_sub_agent_cmd_tests.rs (#[path]) to stay under the 8000-token gate.

Invariant: a pinned subtree stays pinned at every depth; consume + propagate
are now uniform across run/review/debate/batch/plan/claude_sub_agent.

Tests (run_cmd_model_pin_tests.rs, +5): reviewer/debater child env carries the
pin; unpinned review/debate injects nothing; propagate_inherited_subtree_pin
cascades at CSA_DEPTH>0, no-ops at depth 0 and when unpinned. Round-1/2/3 stay
green. Addresses the round-4 review finding on #1741.
