+++ crates/cli-sub-agent/src/batch.rs
@@ -558,10 +558,15 @@ async fn execute_task(
             };
         }
     };
-    let extra_env = global_config.build_execution_env(
+    let mut extra_env = global_config.build_execution_env(
         executor.tool_name(),
         csa_config::ExecutionEnvOptions::default(),
     );
+    // #1741: a batch task picks its own tool/model from the batch TOML and does
+    // NOT consume the parent's subtree pin for that choice, but it MUST still
+    // cascade an inherited pin so nested CSA calls from the task stay pinned all
+    // the way down. Pass-through (no-op unless this process is a pinned child).
+    crate::run_cmd_model_pin::propagate_inherited_subtree_pin(&mut extra_env);
     let extra_env_ref = extra_env.as_ref();
     let idle_timeout_seconds = crate::pipeline::resolve_idle_timeout_seconds(config, None);
     let initial_response_timeout_seconds =
